refactor: move removeCowFromInventory to its own file
